separate date picker input from activity screen
